+++ CHANGELOG.md
@@ -26,6 +26,7 @@ This update accompanies the release of the [Telegram Bot API 6.7](https://core.t
    to `\TgBotLib\Objects\Telegram\InlineKeyboardButton`
  * Added methods `\TgBotLib\Objects\Telegram\InlineKeyboardMarkup::removeRow` & `\TgBotLib\Objects\Telegram\InlineKeyboardMarkup::addRow` to `\TgBotLib\Objects\Telegram\InlineKeyboardMarkup`
  * Added method `\TgBotLib\Objects\Telegram\WebAppInfo::setUrl` to `\TgBotLib\Objects\Telegram\WebAppInfo`
+ * Added methods `\TgBotLib\Classes\Validate::url`, `\TgBotLib\Classes\Validate::length`, & `\TgBotLib\Classes\Validate::isHttps` to `\TgBotLib\Classes\Validate`
 
 ## [6.6.0] - 2023-04-10